✨ Only specify task if default task is sentence similiraty & requested task is feature extraction, or vice versa
