+++ packages/inference/src/lib/isUrl.ts
@@ -1,0 +1,3 @@
+export function isUrl(modelOrUrl: string): boolean {
+	return /^http(s?):/.test(modelOrUrl) || modelOrUrl.startsWith("/");
+}
